refactor: :recycle: change directory structure
stores & types moved into utils/ , layouts & guards moved into components
