+++ app/src/assets/icons/HelpIconSvg.tsx
@@ -1,4 +1,4 @@
-import { MenuIconSvgProps } from '@/types/Menu';
+import { MenuIconSvgProps } from '@/utils/types/Menu';
 import { useTheme } from '@emotion/react';
 
 export const HelpIconSvg = ({ isFocused }: MenuIconSvgProps) => {
