+++ app/src/assets/icons/SettingsIconSvg.tsx
@@ -1,4 +1,4 @@
-import { MenuIconSvgProps } from '@/types/Menu';
+import { MenuIconSvgProps } from '@/utils/types/Menu';
 import { useTheme } from '@emotion/react';
 
 export const SettingsIconSvg = ({ isFocused }: MenuIconSvgProps) => {
