+++ app/src/components/elements/Sidebar/Logo.tsx
